part of final separate avatar commit
Before, as a starting part, there was only one avatar variable, which caused a duplication of animations and ball handling was a problem. now, this problem doesnt happen because theres a local and remote avatar variable. second part, will be an update on server.js which sends the animation variable (the animation the remote variable is going through).